go/consensus/tendermint: Move GenesisProvider to api
