+++ .changelog/3115.internal.1.md
@@ -1,0 +1,1 @@
+go/consensus/tendermint: Move GenesisProvider to api package
